requirements.txt
flask
flask-cors
python-dotenv
supabase
postgrest
psycopg2-binary
gunicorn
